requirements.txt
# HTTP requests and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0  # For dynamic websites if needed

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# String matching and fuzzy search
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# LLM integration
openai>=1.3.0  # For OpenAI GPT models
anthropic>=0.7.0  # Alternative: Claude API

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0  # Progress bars
colorama>=0.4.6  # Colored terminal output
python-dateutil>=2.8.0

# Logging
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
